cleanups from Sam: format, spl's, ptrace defs
SCCS-vsn: sys/kern/kern_acct.c 6.9
SCCS-vsn: sys/kern/kern_fork.c 6.8
SCCS-vsn: sys/kern/sys_generic.c 6.14
SCCS-vsn: sys/kern/sys_process.c 6.6
SCCS-vsn: sys/kern/tty_subr.c 6.11

+++ usr/src/sys/kern/kern_acct.c
@@ -3,7 +3,7 @@
  * All rights reserved.  The Berkeley software License Agreement
  * specifies the terms and conditions for redistribution.
  *
- *	@(#)kern_acct.c	6.8 (Berkeley) %G%
+ *	@(#)kern_acct.c	6.9 (Berkeley) %G%
  */
 
 #include "param.h"
@@ -49,9 +49,9 @@ sysacct()
 		ndp->ni_segflg = UIO_USERSPACE;
 		ndp->ni_dirp = uap->fname;
 		ip = namei(ndp);
-		if(ip == NULL)
+		if (ip == NULL)
 			return;
-		if((ip->i_mode & IFMT) != IFREG) {
+		if ((ip->i_mode&IFMT) != IFREG) {
 			u.u_error = EACCES;
 			iput(ip);
 			return;
@@ -122,7 +122,7 @@ acct()
 	else
 		ap->ac_mem = 0;
 	ap->ac_mem >>= CLSIZELOG2;
-	ap->ac_io = compress(ru->ru_inblock + ru->ru_oublock, 0);
+	ap->ac_io = compress(ru->ru_inblock + ru->ru_oublock, (long)0);
 	if (u.u_ttyp)
 		ap->ac_tty = u.u_ttyd;
 	else
